automate node ensure maintenance on before trying to patch chef server
