+++ components/automate-cli/cmd/chef-automate/automateConfigUtils.go
@@ -139,49 +139,3 @@ func validateIPs(templateCertIPs []IP, infraIPs []string, serviceName string) []
 	}
 	return ipValidationErrors
 }
-
-func ipContain(ips []IP, ip string) bool {
-	for _, i := range ips {
-		if i.IP == ip {
-			return true
-		}
-	}
-	return false
-}
-
-func validateAllServiceIps(templateCerts *CertificateToml, infra *AutomateHAInfraDetails) []error {
-	combinedErrors := []error{}
-
-	if err := validateIPs(
-		templateCerts.OpenSearch.IPS,
-		infra.Outputs.OpensearchPrivateIps.Value,
-		"Opensearch",
-	); err != nil {
-		combinedErrors = append(combinedErrors, err...)
-	}
-	if err := validateIPs(
-		templateCerts.PostgreSQL.IPS,
-		infra.Outputs.PostgresqlPrivateIps.Value,
-		"PostgreSQL",
-	); err != nil {
-		combinedErrors = append(combinedErrors, err...)
-	}
-
-	if err := validateIPs(
-		templateCerts.Automate.IPS,
-		infra.Outputs.AutomatePrivateIps.Value,
-		"Automate",
-	); err != nil {
-		combinedErrors = append(combinedErrors, err...)
-	}
-
-	if err := validateIPs(
-		templateCerts.ChefServer.IPS,
-		infra.Outputs.ChefServerPrivateIps.Value,
-		"ChefServer",
-	); err != nil {
-		combinedErrors = append(combinedErrors, err...)
-	}
-
-	return combinedErrors
-}
